API/leaf-api/requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.16.1
keras==3.3.3
numpy
pillow
python-multipart